(B) BEL-1335 Allow the worker command to be overridden.
So that we can use workers other than sidekiq, allow the worker command to be overridden.

This will mean we can use inst-jobs for canvas.
